design: board UI 수정
