--- frontend/src/components/pages/board/Board.scss
@@ -1,77 +1,0 @@
-.bgBoard {
-  display: flex;
-  flex-direction: column;
-  align-items: center; /* 수직 가운데 정렬 */
-  text-align: center; /* 가로 가운데 정렬 */
-  background-color: aliceblue;
-  justify-content: center;
-}
-
-.container {
-  width: 1200px;
-  position: relative;
-  border: 1px solid black;
-}
-
-.signinBtn {
-  position: absolute;
-  top: 10px;
-  right: 0;
-  margin: 10px;
-}
-
-.logo {
-  // border: 1px solid black;
-  margin: 40px;
-  font-size: 35px;
-}
-
-.nav {
-  width: 1200px;
-  // padding: 10px;
-  // margin: 10px;
-  background-color: pink;
-  display: flex;
-  justify-content: space-around;
-}
-
-.category {
-  border: 1px solid black;
-  width: 600px;
-  display: flex;
-  padding: 0px;
-  margin: 10px;
-  justify-content:space-around
-}
-
-
-.contents {
-  width: 1200px;
-  display: flex;
-  justify-content:center;
-
-  background-color: white;
-}
-
-.box {
-  width: 1200px;
-}
-
-.bar {  
-  width: 574px;
-  height: 220px;
-  padding: 10px;
-  float:left;
-
-  border: 1px solid black;
-}
-
-
-
-.writeButton {
-  width: 100px;
-  margin: 50px;
-  padding: 10px;
-
-  border: solid 1px red;
-}
